fake_id0: Fix POKE_MEM_ID to call poke_uint32 instead of poke_uint16
Fix getresuid() and getresgid() uses the POKE_MEM_ID macro when using
fake_id0, this was the only macro that was using 16bits return value
for IDs.

This behavior causes some applications, such as apt-get, to not get
proper results when the host user has an UID greater than 16bits.